backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.49
alembic>=1.13.1
pydantic>=2.7.0
pydantic-settings>=2.2.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
aiosqlite==0.19.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.23.1
python-dotenv==1.0.0
